3.2-crud/stocks_products/requirements.txt
django
django-filter==23.1
djangorestframework
psycopg2-binary
